requirements.txt
pytest == 7.2.0
allure-pytest
selene == 2.0.0b16